requirements.txt
PyJWT==2.10.1
requests==2.32.3
gql==3.5.1
aiohttp==3.11.13
